+++ src/YamNet.Client35/Client.cs
@@ -9,8 +9,6 @@ namespace YamNet.Client
     using System;
     using System.Net;
 
-//    using YamNet.Client.Clients;
-
     /// <summary>
     /// The Yammer web client.
     /// </summary>